Fixed owner rights

+++ main.py
@@ -122,7 +122,7 @@ async def strictness_command(message: Message):
     member = await bot.get_chat_member(message.chat.id, message.from_user.id)
     logger.info(f'Admin: {member.status}')
     logger.info(f'Can delete messages: {member.can_delete_messages}')
-    if member.status != ChatMemberStatus.ADMINISTRATOR or not member.can_delete_messages:
+    if member.status == ChatMemberStatus.CREATOR and (member.status not in ChatMemberStatus.ADMINISTRATOR or not member.can_delete_messages):
         tmsg = await message.reply('Это может настраивать только админ с правами на удаление сообщений.')
         await delete_after(message, tmsg)
         return
